refactor(festival): move previous website to 2020 folder
Move previous website to 2020 folder

BREAKING CHANGE: All links are now /2020/[link]
