+++ src/pages/2020/speakers.js
@@ -1,12 +1,12 @@
 import React from "react";
 
-import Layout from "../components/layout";
-import SEO from "../components/SEO/seo";
+import Layout from "../../components/layout";
+import SEO from "../../components/SEO/seo";
 
-import Nav from "../components/nav";
-import Breadcrumb from "../components/breadcrumb";
-import Sched from "../components/schedSpeakers";
-import Footer from "../components/footer";
+import Nav from "../../components/nav";
+import Breadcrumb from "../../components/breadcrumb";
+import Sched from "../../components/schedSpeakers";
+import Footer from "../../components/footer";
 
 function Speakers() {
   return (
