+++ src/pages/2020/sustain-africa.js
@@ -1,11 +1,11 @@
 import React from "react";
 
-import Layout from "../components/layout";
-import SEO from "../components/SEO/seo";
+import Layout from "../../components/layout";
+import SEO from "../../components/SEO/seo";
 
-import Nav from "../components/nav";
-import SustainHead from "../components/sustainHead";
-import Footer from "../components/footer";
+import Nav from "../../components/nav";
+import SustainHead from "../../components/sustainHead";
+import Footer from "../../components/footer";
 
 function SustainAfrica() {
   return (
