+++ src/pages/index.js
@@ -1,6 +1,5 @@
 import React from "react";
 
-import Layout from "../components/layout";
 import SEO from "../components/SEO/seo";
 
 import Nav from "../components/nav";
@@ -14,7 +13,7 @@ import Footer from "../components/footer";
 
 function IndexPage() {
   return (
-    <Layout>
+    <>
       <SEO keywords={["Open Source Community Africa", "OSCA", "Open Source Festival", "OSF"]} title="Home" />
       <Nav />
       <Header />
@@ -24,7 +23,7 @@ function IndexPage() {
       <Sponsors />
       <Gallery />
       <Footer />
-    </Layout>
+    </>
   );
 }
